requirements.txt
fastapi
uvicorn
pyjwt
pydantic